fix(wsl): 🐛 uname no longer references wsl
Prior to this change, to discover wsl environments, the specific phrase
"WSL" should be included in the when executing uname -a. However it
seems like in recent releases WSL no longer exists.

running `uname -a` on a new minted wsl ubuntu installation returned the
following

```
Linux ERMLAPTOP902 4.4.0-19041-Microsoft #1237-Microsoft Sat Sep 11 14:32:00 PST 2021 x86_64 x86_64 x86_64 GNU/Linux
```

After this change, the discover function is updated to search for
Microservice as well as WSL to remain for backwards compatibility.