fix(aerospace): 🐛 move focus while moving node
Previously, when you moved a application or window to a different
monitor, the focus would not move to the new monitor. This would mean
that you would have to manually focus on the new monitor, which is
counterintuitive - as the focus should move to the new monitor so that I
can continue moving the window around (this is coupled by the fact that
I am using --wrap-around).

This commit fixes the issue by adding a focus-monitor command to the
move around.

Signed-off-by: Sriram Venkatesh <sriram.venkatesh@versent.com.au>