fix(aerospace,ghostty): 🐛 render windows as seperate tabs
Not entierly sure why this is needed (even though ghostty documents it),
but I was having issues with floating windows not being resized
correctly by aerospace.

I thought this was the fix, but it didn't work - so have just left it as
is.

The floating window thing fixed it self, after a restart.. so not sure
what happened there.

https://ghostty.org/docs/help/macos-tiling-wms

Signed-off-by: Sriram Venkatesh <sriram.venkatesh@versent.com.au>